+++ src/perlin.h
@@ -11,17 +11,17 @@ public:
 	Perlin();
 	~Perlin();
 
-float noise3D(float x, float y, float z);
-float fade(float t);
-float lerp(float t, float a, float b);
-float grad(int hash, float x, float y, float z);
+double noise3D(double x, double y, double z);
+double fade(double t);
+double lerp(double t, double a, double b);
+double grad(int hash, double x, double y, double z);
 
 
 
 
 
 private:
-	std::vector<float> p = {1.0, 2.0, 3.0};
+	std::vector<double> p = {1.0, 2.0, 3.0};
 };
 
 #endif
